Exercise Recommendation is 90% finished. Just need to add in a way for the user to be able to store it in a table.

+++ BetterU/NutritionDataFromBarcodeViewController.swift
@@ -235,9 +235,9 @@ class NutritionDataFromBarcodeViewController: UIViewController, UITableViewDeleg
             self.servingSizeUnit = servingSizeUnit
         }
         
-        if let servingSizeQty = jsonDict["nf_serving_size_qty"] as? Int
+        if let servingSizeQty = jsonDict["nf_serving_size_qty"] as? Double
         {
-            self.servingSizeQty = String(Double(servingSizeQty))
+            self.servingSizeQty = String(servingSizeQty)
         }
         
         if let name = jsonDict["item_name"] as? String {
